move message parsing from Helper into FinTsClient
