+++ src/libfintx.FinTS/Segments/HKCCS.cs
@@ -119,7 +119,7 @@ namespace libfintx.FinTS
             var message = FinTSMessage.Create(client, client.HNHBS, client.HNHBK, segments, client.HIRMS);
             var response = await FinTSMessage.Send(client, message);
 
-            Helper.Parse_Message(client, response);
+            client.Parse_Message(response);
 
             return response;
         }
