+++ src/libfintx.FinTS/Segments/HKCDE.cs
@@ -82,7 +82,7 @@ namespace libfintx.FinTS
             string message = FinTSMessage.Create(client, client.HNHBS, client.HNHBK, segments, client.HIRMS);
             var response = await FinTSMessage.Send(client, message);
 
-            Helper.Parse_Message(client, response);
+            client.Parse_Message(response);
 
             return response;
         }
