+++ src/libfintx.FinTS/Segments/HKCME.cs
@@ -80,7 +80,7 @@ namespace libfintx.FinTS
             string message = FinTSMessage.Create(client, client.HNHBS, client.HNHBK, segments, client.HIRMS);
             var response = await FinTSMessage.Send(client, message);
 
-            Helper.Parse_Message(client, response);
+            client.Parse_Message(response);
 
             return response;
         }
