+++ src/libfintx.FinTS/Segments/HKCUM.cs
@@ -75,7 +75,7 @@ namespace libfintx.FinTS
             var TAN = await FinTSMessage.Send(client, FinTSMessage.Create(client, client.HNHBS, client.HNHBK, segments,
                 client.HIRMS));
 
-            Helper.Parse_Message(client, TAN);
+            client.Parse_Message(TAN);
 
             return TAN;
         }
